+++ lib/src/impl/real_storage.dart
@@ -1,31 +1,39 @@
 import 'dart:async';
 import 'dart:collection';
 
+import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
+import 'package:rxdart_ext/rxdart_ext.dart';
 
+import '../async_memoizer.dart';
 import '../interface/rx_storage.dart';
 import '../interface/storage.dart';
 import '../logger/logger.dart';
 import '../model/key_and_value.dart';
-import '../stream_extensions/map_not_null.dart';
-import '../stream_extensions/single_subscription.dart';
 
-/// Default [RxStorage] implementation
-class RealRxStorage implements RxStorage {
+// ignore_for_file: unnecessary_null_comparison
+
+/// Default [RxStorage] implementation.
+class RealRxStorage<Key extends Object, Options,
+    S extends Storage<Key, Options>> implements RxStorage<Key, Options> {
   static const _initialKeyValue =
       KeyAndValue('RealRxStorage', 'hoc081098@gmail.com');
 
   /// Trigger subject
-  final _keyValuesSubject = PublishSubject<Map<String, Object?>>();
+  final _keyValuesSubject = PublishSubject<Map<Key, Object?>>();
+
+  final _disposeMemo = AsyncMemoizer<void>();
+
+  var _isDisposed = false;
 
   /// Logger subscription. Nullable
-  StreamSubscription<Map<String, dynamic>>? _subscription;
+  StreamSubscription<Map<Key, Object?>>? _subscription;
 
   /// Nullable.
-  Storage? _storage;
+  S? _storage;
 
   /// Nullable.
-  Future<Storage>? _storageFuture;
+  late Future<S> _storageFuture;
 
   /// Nullable
   final Logger? _logger;
@@ -35,181 +43,99 @@ class RealRxStorage implements RxStorage {
 
   /// Construct a [RealRxStorage].
   RealRxStorage(
-    FutureOr<Storage> storageOrFuture, [
+    FutureOr<S> storageOrFuture, [
     this._logger,
     this._onDispose,
-  ]) {
-    if (storageOrFuture is Future<Storage>) {
+  ]) : assert(storageOrFuture != null) {
+    if (storageOrFuture is Future<S>) {
       _storageFuture = storageOrFuture.then((value) => _storage = value);
     } else {
-      _storageFuture = null;
       _storage = storageOrFuture;
     }
 
-    _subscription = _logger?.let(
-      (logger) => _keyValuesSubject.listen((map) {
+    _subscription = _logger?.let((logger) {
+      _keyValuesSubject.listen((map) {
         final pairs = [
           for (final entry in map.entries)
-            KeyAndValue(
+            KeyAndValue<Key, Object?>(
               entry.key,
               entry.value,
             ),
         ];
         logger.keysChanged(UnmodifiableListView(pairs));
-      }),
-    );
+      });
+    });
   }
 
   //
   // Internal
   //
 
-  /// Workaround to capture generics
-  static Type _typeOf<T>() => T;
-
-  /// Read value from persistent [storage] by [key].
-  static Future<dynamic> _readFromStorage<T>(Storage storage, String key) {
-    if (T == dynamic) {
-      return storage.get(key);
-    }
-    if (T == double) {
-      return storage.getDouble(key);
-    }
-    if (T == int) {
-      return storage.getInt(key);
-    }
-    if (T == bool) {
-      return storage.getBool(key);
-    }
-    if (T == String) {
-      return storage.getString(key);
-    }
-    if (T == _typeOf<List<String>>()) {
-      return storage.getStringList(key);
-    }
-    throw StateError('Unhandled type $T');
+  bool _debugAssertNotDisposed() {
+    assert(() {
+      if (_isDisposed && _disposeMemo.hasRun) {
+        throw StateError('A $runtimeType was used after being disposed.\n'
+            'Once you have called dispose() on a $runtimeType, it can no longer be used.');
+      }
+      return true;
+    }());
+    return true;
   }
 
-  /// Write [value] to [storage] associated with [key]
-  static Future<bool> _writeToStorage<T>(
-    Storage storage,
-    String key,
-    T? value,
-  ) {
-    if (T == dynamic) {
-      assert(value == null);
-      return storage.remove(key);
-    }
-
-    final dynamicVal = value as dynamic;
-    if (T == double) {
-      return storage.setDouble(key, dynamicVal);
-    }
-    if (T == int) {
-      return storage.setInt(key, dynamicVal);
-    }
-    if (T == bool) {
-      return storage.setBool(key, dynamicVal);
-    }
-    if (T == String) {
-      return storage.setString(key, dynamicVal);
-    }
-    if (T == _typeOf<List<String>>()) {
-      return storage.setStringList(key, dynamicVal);
-    }
-
-    throw StateError('Unhandled type $T');
-  }
-
-  /// Get [Stream] from the persistent storage
-  Stream<T?> _getStream<T>(String key) {
-    final stream = _keyValuesSubject
-        .toSingleSubscriptionStream()
-        .mapNotNull(
-            (map) => map.containsKey(key) ? KeyAndValue(key, map[key]) : null)
-        .startWith(_initialKeyValue) // Dummy value to trigger initial load.
-        .asyncMap<T?>((entry) => identical(_initialKeyValue, entry)
-            ? _getValue<T>(key)
-            : entry.value as FutureOr<T?>);
-
-    return _logger?.let((logger) => stream
-            .doOnData((value) => logger.doOnDataStream(KeyAndValue(key, value)))
-            .doOnError(
-                (e, s) => logger.doOnErrorStream(e, s ?? StackTrace.empty))) ??
-        stream;
-  }
-
-  /// Get value from the persistent [Storage] by [key].
-  Future<T?> _getValue<T>(String key) async {
-    final storage = _storage ?? await _storageFuture!;
-    final T? value = await _readFromStorage<T>(storage, key);
-
-    _logger?.readValue(T, key, value);
-    return value;
-  }
-
-  /// Set [value] associated with [key].
-  Future<bool> _setValue<T>(String key, T? value) async {
-    final storage = _storage ?? await _storageFuture!;
-    final result = await _writeToStorage<T>(storage, key, value);
-
-    _logger?.writeValue(T, key, value, result);
-    if (result) {
-      _sendKeyValueChanged({key: value});
-    }
-
-    return result;
-  }
-
-  /// Add pairs to subject to trigger.
-  /// Do nothing if subject already closed.
-  void _sendKeyValueChanged(Map<String, dynamic> map) {
+  /// Add changed map to subject to trigger.
+  @protected
+  void sendChange(Map<Key, Object?> map) {
     try {
       _keyValuesSubject.add(map);
-    } catch (e) {
-      print(e);
-      // Do nothing
+    } on StateError {
+      assert(_debugAssertNotDisposed());
     }
   }
+
+  /// Calling [block] with [S] as argument.
+  @protected
+  Future<R> useStorage<R>(Future<R> Function(S) block) =>
+      _storage?.let(block) ?? _storageFuture.then(block);
 
   // Get and set methods (implements [Storage])
 
   @override
-  Future<bool> containsKey(String key) async {
-    final storage = _storage ?? await _storageFuture!;
-    return storage.containsKey(key);
+  Future<bool> containsKey(Key key, [Options? options]) async {
+    assert(_debugAssertNotDisposed());
+    assert(key != null);
+
+    return useStorage((s) => s.containsKey(key, options));
   }
 
   @override
-  Future<Object?> get(String key) => _getValue<dynamic>(key);
+  Future<T?> read<T extends Object>(Key key, Decoder<T?> decoder,
+      [Options? options]) async {
+    assert(_debugAssertNotDisposed());
+    assert(key != null);
+    assert(decoder != null);
 
-  @override
-  Future<bool?> getBool(String key) => _getValue<bool>(key);
-
-  @override
-  Future<double?> getDouble(String key) => _getValue<double>(key);
-
-  @override
-  Future<int?> getInt(String key) => _getValue<int>(key);
-
-  @override
-  Future<Set<String>> getKeys() async {
-    final storage = _storage ?? await _storageFuture!;
-    return storage.getKeys();
+    final value = await useStorage((s) => s.read(key, decoder, options));
+    _logger?.readValue(T, key, value);
+    return value;
   }
 
   @override
-  Future<String?> getString(String key) => _getValue<String>(key);
+  Future<Map<Key, Object?>> readAll([Options? options]) async {
+    assert(_debugAssertNotDisposed());
+
+    final all = await useStorage((s) => s.readAll(options));
+    _logger?.let((logger) {
+      all.forEach((key, value) => logger.readValue(dynamic, key, value));
+    });
+    return all;
+  }
 
   @override
-  Future<List<String>?> getStringList(String key) =>
-      _getValue<List<String>>(key);
+  Future<bool> clear([Options? options]) async {
+    assert(_debugAssertNotDisposed());
 
-  @override
-  Future<bool> clear() async {
-    final storage = _storage ?? await _storageFuture!;
-    final keys = await storage.getKeys();
-    final result = await storage.clear();
+    final keys = (await readAll()).keys;
+    final result = await useStorage((s) => s.clear(options));
 
     // All values are set to null
     _logger?.let((logger) {
@@ -219,90 +145,103 @@ class RealRxStorage implements RxStorage {
     });
     if (result) {
       final map = {for (final k in keys) k: null};
-      _sendKeyValueChanged(map);
+      sendChange(map);
     }
 
     return result;
   }
 
   @override
-  Future<void> reload() async {
-    final storage = _storage ?? await _storageFuture!;
-    await storage.reload();
+  Future<bool> remove(Key key, [Options? options]) async {
+    assert(_debugAssertNotDisposed());
+    assert(key != null);
 
-    final keys = await storage.getKeys();
+    final result = await useStorage((s) => s.remove(key, options));
 
-    // Read new values from storage.
-    final map = {for (final k in keys) k: await storage.get(k)};
-    _logger?.let((logger) {
-      for (final key in keys) {
-        logger.readValue(dynamic, key, map[key]);
-      }
-    });
-    _sendKeyValueChanged(map);
+    _logger?.writeValue(dynamic, key, null, result);
+    if (result) {
+      sendChange({key: null});
+    }
+
+    return result;
   }
 
   @override
-  Future<bool> remove(String key) => _setValue<dynamic>(key, null);
+  Future<bool> write<T extends Object>(Key key, T? value, Encoder<T?> encoder,
+      [Options? options]) async {
+    assert(_debugAssertNotDisposed());
+    assert(key != null);
+    assert(encoder != null);
 
-  @override
-  Future<bool> setBool(String key, bool? value) => _setValue<bool>(key, value);
+    final result =
+        await useStorage((s) => s.write(key, value, encoder, options));
 
-  @override
-  Future<bool> setDouble(String key, double? value) =>
-      _setValue<double>(key, value);
+    _logger?.writeValue(T, key, value, result);
+    if (result) {
+      sendChange({key: value});
+    }
 
-  @override
-  Future<bool> setInt(String key, int? value) => _setValue<int>(key, value);
-
-  @override
-  Future<bool> setString(String key, String? value) =>
-      _setValue<String>(key, value);
-
-  @override
-  Future<bool> setStringList(String key, List<String>? value) =>
-      _setValue<List<String>>(key, value);
+    return result;
+  }
 
   // Get streams (implements [RxStorage])
 
   @override
-  Stream<Object?> getStream(String key) => _getStream<dynamic>(key);
+  Stream<T?> observe<T extends Object>(Key key, Decoder<T?> decoder,
+      [Options? options]) {
+    assert(_debugAssertNotDisposed());
+    assert(key != null);
+
+    final stream = _keyValuesSubject
+        .toSingleSubscriptionStream()
+        .mapNotNull((map) => map.containsKey(key)
+            ? KeyAndValue<Object, Object?>(key, map[key])
+            : null)
+        .startWith(_initialKeyValue) // Dummy value to trigger initial load.
+        .asyncMap<T?>(
+          (entry) => identical(entry, _initialKeyValue)
+              ? read<T>(key, decoder, options)
+              : entry.value as FutureOr<T?>,
+        );
+
+    return _logger?.let((logger) {
+          return stream
+              .doOnData(
+                  (value) => logger.doOnDataStream(KeyAndValue(key, value)))
+              .doOnError(
+                  (e, s) => logger.doOnErrorStream(e, s ?? StackTrace.current));
+        }) ??
+        stream;
+  }
 
   @override
-  Stream<bool?> getBoolStream(String key) => _getStream<bool>(key);
+  Stream<Map<Key, Object?>> observeAll([Options? options]) {
+    assert(_debugAssertNotDisposed());
+
+    return _keyValuesSubject
+        .toSingleSubscriptionStream()
+        .startWith(const {}).asyncMap((_) => readAll(options));
+  }
 
   @override
-  Stream<double?> getDoubleStream(String key) => _getStream<double>(key);
+  Future<void> dispose() {
+    assert(_debugAssertNotDisposed());
 
-  @override
-  Stream<int?> getIntStream(String key) => _getStream<int>(key);
+    return _disposeMemo.runOnce(() async {
+      final cancelFuture = _subscription?.cancel();
 
-  @override
-  Stream<String?> getStringStream(String key) => _getStream<String>(key);
+      if (cancelFuture == null) {
+        await _keyValuesSubject.close();
+      } else {
+        await Future.wait(
+          [_keyValuesSubject.close(), cancelFuture],
+          eagerError: true,
+        );
+      }
 
-  @override
-  Stream<List<String>?> getStringListStream(String key) =>
-      _getStream<List<String>>(key);
-
-  @override
-  Stream<Set<String>> getKeysStream() => _keyValuesSubject
-      .toSingleSubscriptionStream()
-      .startWith(const <String, dynamic>{}).asyncMap((_) => getKeys());
-
-  @override
-  Future<void> dispose() async {
-    final cancelFuture = _subscription?.cancel();
-
-    if (cancelFuture == null) {
-      await _keyValuesSubject.close();
-    } else {
-      await Future.wait(
-        [_keyValuesSubject.close(), cancelFuture],
-        eagerError: true,
-      );
-    }
-
-    _onDispose?.call();
+      _isDisposed = true;
+      _onDispose?.call();
+    });
   }
 }