+++ test/fake_storage.dart
@@ -1,18 +1,29 @@
+import 'dart:async';
+
+import 'package:rx_storage/rx_storage.dart';
 import 'package:rx_storage/src/interface/storage.dart';
 
 import 'utils/synchronous_future.dart';
 
 Future<T> _wrap<T>(T value) => SynchronousFuture(value);
 
-class FakeStorage implements Storage {
-  Map<String, dynamic> _map;
-  Map<String, dynamic>? _pendingMap;
+abstract class StringKeyStorage extends Storage<String, void> {
+  Future<void> reload();
+}
 
-  FakeStorage(Map<String, dynamic> map) : _map = Map.of(map);
+abstract class StringKeyRxStorage extends StringKeyStorage
+    implements RxStorage<String, void> {}
 
-  set map(Map<String, dynamic> map) => _pendingMap = Map.of(map);
+class FakeStorage implements StringKeyStorage {
+  Map<String, Object?> _map;
+  Map<String, Object?>? _pendingMap;
 
-  Future<bool> _setValue(String key, dynamic value) {
+  FakeStorage(Map<String, Object?> map) : _map = Map<String, Object?>.of(map);
+
+  set map(Map<String, Object?> map) =>
+      _pendingMap = Map<String, Object?>.of(map);
+
+  Future<bool> _setValue(String key, Object? value) {
     if (value is List<String>?) {
       _map[key] = value?.toList();
     } else {
@@ -23,9 +34,7 @@ class FakeStorage implements Storage {
 
   Future<T?> _getValue<T>(String key) {
     final value = _map[key] as T?;
-    return value is List<String>
-        ? _wrap(value.toList() as dynamic)
-        : _wrap(value);
+    return value is List<String> ? _wrap(value.toList() as T) : _wrap(value);
   }
 
   //
@@ -33,34 +42,20 @@ class FakeStorage implements Storage {
   //
 
   @override
-  Future<bool> clear() {
+  Future<bool> clear([void _]) {
     _map.clear();
     return _wrap(true);
   }
 
   @override
-  Future<bool> containsKey(String key) => _wrap(_map.containsKey(key));
+  Future<bool> containsKey(String key, [void _]) =>
+      _wrap(_map.containsKey(key));
 
   @override
-  Future<Object?> get(String key) => _getValue(key);
-
-  @override
-  Future<bool?> getBool(String key) => _getValue(key);
-
-  @override
-  Future<double?> getDouble(String key) => _getValue(key);
-
-  @override
-  Future<int?> getInt(String key) => _getValue(key);
-
-  @override
-  Future<Set<String>> getKeys() => _wrap(_map.keys.toSet());
-
-  @override
-  Future<String?> getString(String key) => _getValue(key);
-
-  @override
-  Future<List<String>?> getStringList(String key) => _getValue(key);
+  Future<bool> write<T extends Object>(
+          String key, T? value, Encoder<T?> encoder,
+          [void _]) =>
+      _setValue(key, encoder(value));
 
   @override
   Future<void> reload() {
@@ -72,21 +67,33 @@ class FakeStorage implements Storage {
   }
 
   @override
-  Future<bool> remove(String key) => _setValue(key, null);
+  Future<bool> remove(String key, [void _]) => _setValue(key, null);
 
   @override
-  Future<bool> setBool(String key, bool? value) => _setValue(key, value);
+  Future<T?> read<T extends Object>(String key, Decoder<T?> decoder,
+          [void _]) =>
+      _getValue<Object>(key).then(decoder);
 
   @override
-  Future<bool> setDouble(String key, double? value) => _setValue(key, value);
+  Future<Map<String, Object?>> readAll([void _]) =>
+      _wrap(<String, Object?>{..._map});
+}
+
+class FakeRxStorage extends RealRxStorage<String, void, StringKeyStorage>
+    implements StringKeyRxStorage {
+  FakeRxStorage(
+    FutureOr<StringKeyStorage> storageOrFuture, [
+    Logger? logger,
+    void Function()? onDispose,
+  ]) : super(
+          storageOrFuture,
+          logger,
+          onDispose,
+        );
 
   @override
-  Future<bool> setInt(String key, int? value) => _setValue(key, value);
-
-  @override
-  Future<bool> setString(String key, String? value) => _setValue(key, value);
-
-  @override
-  Future<bool> setStringList(String key, List<String>? value) =>
-      _setValue(key, value);
+  Future<void> reload() async {
+    await useStorage((s) => s.reload());
+    sendChange(await readAll());
+  }
 }