+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2 - Jan 4, 2021
+
+-   Refactors `Storage` and `RxStorage`:
+    -   Supports any type via `Encoder` and `Decoder`.
+    -   Generic `Key` and generic `Options`.
+-   Exports `RealRxStorage` class.
+
 ## 0.0.1 - Dec 14, 2020
 
 -   Initial version.